4 - Capstone Project/requirements.txt
[venv dependencies]

venv python 3.12.0
django 4.2.7
djangorestframework 3.14.0
mysqlclient 2.0.1